# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API documentation for the Totes backend using Gin and Swagger.",
        "title": "Totes Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@totesbackend.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/additional-expenses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves all additional expense records. Requires permission to view all additional expenses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "additional-expenses"
                ],
                "summary": "Get all additional expenses",
                "responses": {
                    "200": {
                        "description": "A list of all additional expenses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AdditionalExpense"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or permission denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error retrieving additional expenses",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new additional expense record. Requires permission to create an additional expense.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "additional-expenses"
                ],
                "summary": "Create a new additional expense",
                "parameters": [
                    {
                        "description": "Additional Expense DTO",
                        "name": "expense",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateAdditionalExpenseDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The created additional expense",
                        "schema": {
                            "$ref": "#/definitions/models.AdditionalExpense"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or permission denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error creating additional expense",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/additional-expenses/{id}": {
            "get": {
                "description": "Retrieves an additional expense record by its ID. Requires permission to view additional expenses by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "additional-expenses"
                ],
                "summary": "Get additional expense by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the additional expense",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The additional expense record",
                        "schema": {
                            "$ref": "#/definitions/models.AdditionalExpense"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or permission denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Additional expense not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error (log registration or DB error)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the details of an existing additional expense identified by its ID. Requires permission to update an additional expense.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "additional-expenses"
                ],
                "summary": "Update an additional expense by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Additional Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Additional Expense details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateAdditionalExpenseDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated additional expense",
                        "schema": {
                            "$ref": "#/definitions/models.AdditionalExpense"
                        }
                    },
                    "400": {
                        "description": "Invalid request or JSON format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or permission denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Additional expense not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes an additional expense record by its ID. Requires permission to delete an additional expense.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "additional-expenses"
                ],
                "summary": "Delete an additional expense by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Additional Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message indicating successful deletion",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format or request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or permission denied",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Additional expense not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error deleting additional expense",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.UpdateAdditionalExpenseDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "expense": {
                    "type": "number"
                },
                "item_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.AdditionalExpense": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "expense": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}